stream: always fire drain after write if needed
in a writable stream the drain event is fired after write
only if the state is empty, this can lead to a deadlock,
for example, it happens to this writable stream:

1. the writable stream is corked
2. data is written to stream
3. the stream is uncorked and corked again
4. _writev gets executed synchronusly
5. the buffer reached high watermark
6. what writes to the stream finds that write returns false,
so it stops writing waiting for the drain event
7. _writev finishes and calls the callback function
8. the callback function, finds that the state is not empty,
so it doesn't fire the drain event

and whatever is writing to the stream keeps hanging forever

The needsDrain flag is only set when the data length exceeds the high
watermark. So it makes sense that it gets reset after writing to the
underlying resource when the data length is less than the high
watermark. The only problem is that if the remaining data in the stream
is bigger that the high watermark, we still have the same issue

Signed-off-by: Mohamed Elawadi <mohamed@elawadi.net>